refactoring to fix #2; separate agents module; main takes options dict; etc.
